requirements_backup.txt
addict==2.4.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.1.8
cloudpickle==3.1.1
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.2
cuda-bindings==12.9.2
cuda-pathfinder==1.3.0
cuda-python==12.9.2
cudf-cu12==25.8.0
cuml-cu12==25.8.0
cupy-cuda12x==13.6.0
cuvs-cu12==25.8.0
cycler==0.12.1
dash==3.2.0
dask==2025.7.0
dask-cuda==25.8.0
dask-cudf-cu12==25.8.0
dbus-python==1.2.18
decorator==5.2.1
distributed==2025.7.0
distributed-ucxx-cu12==0.45.1
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
fastrlock==0.8.3
Flask==3.1.2
fonttools==4.60.1
fsspec==2025.9.0
idna==3.10
ifaddr==0.2.0
importlib-metadata==6.6.0
ipython==8.37.0
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiss-icp==1.2.2
kiwisolver==1.4.9
laspy==2.5.4
libcudf-cu12==25.8.0
libcuml-cu12==25.8.0
libcuvs-cu12==25.8.0
libkvikio-cu12==25.8.0
libraft-cu12==25.8.0
librmm-cu12==25.8.0
libucx-cu12==1.18.1
libucxx-cu12==0.45.1
llvmlite==0.44.0
locket==1.0.0
lz4==4.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.8.0
msgpack==1.1.2
narwhals==2.7.0
natsort==8.4.0
nbformat==5.10.4
nest-asyncio==1.6.0
numba==0.61.2
numba-cuda==0.14.1
numpy==1.26.4
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cccl-cu12==12.9.27
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cufft-cu12==11.4.1.4
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.28.3
nvidia-nvjitlink-cu12==12.9.86
nvtx==0.2.13
open3d==0.18.0
ouster_sdk==0.15.1
packaging==25.0
pandas==2.3.3
parso==0.8.5
partd==1.4.2
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.1
prettytable==3.16.0
prompt_toolkit==3.0.52
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
Pygments==2.19.2
PyGObject==3.42.1
pylibcudf-cu12==25.8.0
pylibraft-cu12==25.8.0
pynvjitlink-cu12==0.7.0
pynvml==12.0.0
pyparsing==3.2.5
pyproj==3.7.1
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
raft-dask-cu12==25.8.0
rapids-dask-dependency==25.8.0
rapids-logger==0.1.1
referencing==0.36.2
requests==2.32.5
retrying==1.4.2
rich==14.1.0
rmm-cu12==25.8.0
rosbags==0.9.23
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
scikit-learn==1.7.2
scipy==1.13.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
stack-data==0.6.3
tblib==3.1.0
threadpoolctl==3.6.0
toolz==1.0.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
treelite==4.4.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ucx-py-cu12==0.45.0
ucxx-cu12==0.45.1
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.3
widgetsnbextension==4.0.14
zeroconf==0.148.0
zict==3.0.0
zipp==3.23.0
zstandard==0.25.0
